0g_py_inference/requirements.txt
# Core dependencies
web3>=6.0.0
eth-account>=0.10.0
eth-utils>=2.0.0
requests>=2.31.0

# Optional dependencies
python-dotenv>=1.0.0

# Optional: required for fine-tuning model decryption
# eciespy>=0.4.0
# cryptography>=41.0.0

# Type hints support (for Python < 3.10)
typing-extensions>=4.0.0